// src/app/models/translate.model.ts
export interface Translation {
  AUTH: Auth;
  BUTTONS: Buttons;
  MESSAGES: Messages;
  LOGGER: Logger;
  STRINGS: Strings;
  SITETITLE: SiteTitle;
  STATUS: Status;
  TABLEHEADINGS: TableHeadings;
  CONFIRMATION: Confirmation;
}

export interface Strings {
  SENT_REQUEST: string;
  USERNAME: string;
  FIRSTNAME: string;
  LASTNAME: string;
  EMAIL: string;
  MOBILE: string;
  LIGHTDEFAULT: string;
  EVILDARK: string;
  OLIVEGREEN: string;
  DARKFOREST: string;
  OTHERSETTINGS: string;
  CHANGETHEME: string;
  EXPAND: string;
  REMOVE: string;
  ADD: string;
  USERINFO: string;
  SITEADDRESS: string;
  SITESAFETYPLAN: string;
  SITESAFETYMANAGER: string;
  SAFEZONE: string;
  SITEINFORMATION: string;
  SITE: string;
  RISK: string;
  RESOLVED: string;
  DATE_TIME: string;
  RESOLVED_BY: string;
  ADDED_BY: string;
  DESCRIPTION: string;
  UPLOAD_DOC: string;
  FILE_NAME: string;
  SELECT_FOLDER: string;
  IMAGE: string;
  VIEW_DOC: string;
  RENAME_FOLDER: string;
  FOLDER_NAME: string;
  CHECK_IN: string;
  CHECK_OUT: string;
  DURATION: string;
  NO_ACTIVITIES: string;
  NO_DOCUMENTS: string;
  NO_FOLDERS: string;
  PASSWORD: string;
  MOBILENO: string;
  PASSWORD1: string;
  PASSWORD2: string;
  NAME: string;
  TYPE: string;
  ADDRESS: string;
  CITY: string;
  COUNTRY: string;
  ZIPCODE: string;
  BILLINGEMAIL: string;
  ACCOUNTNO: string;
  PHONENO: string;
  FAX: string;
  PACKAGE_DESCRIPTION_1: string;
  PACKAGE_DESCRIPTION_2: string;
  PACKAGE_DESCRIPTION_3: string;
  QUESTION_NAME: string;
  QUESTION_WARNING: string;
  SELECT_QUESTION_TYPE: string;
  WHICH_ANSWER_SAFE: string;
  SAFE_QUESTION: string;
  PARENT: string;
  SAFE_QUESTION_YES: string;
  SAFE_QUESTION_NO: string;
  SAFE_QUESTION_BOTH: string;
  CAN_PROCEED_QUESTION: string;
  CAN_PROCEED_QUESTION_YES: string;
  CAN_PROCEED_QUESTION_NO: string;
  RENAME: string;
  DELETE_FOLDER: string;
  PARENT_QUESTION: string;
  CHILD_YES: string;
  CHILD_NO: string;
  CHILD: string;
  CHANGEPICTURE: string;
  ACTION_REPORT: string;
  CHILD_YES_SAFE: string;
  CHILD_NO_SAFE: string;
  SEARCHUSERS: string;
  ADVSEARCH: string;
  CHOOSE_ROLE: string;
  SELECTED_ENTITY: string;
  SELECT_TEAM: string;
  SEARCH: string;
  SELECT_FILTER: string;
  ALLTEAMS: string;
  RANGE: string;
  DATE_TO: string;
  DATE_FROM: string;
  SELECTED_USERS: string;
  MAKE_TEAMLEAD: string;
  REMOVE_TEAMLEAD: string;
  HAZARD: string;
  PRECAUTIONARY_NOTE: string;
  PERSON_FOR_ACTION: string;
  REVIEW: string;
  RENAME_FILE: string;
  ENTITYMANAGER: string;
}

export interface Auth {
  USERNAME: string;
  PASSWORD: string;
  EMAIL: string;
  FIRSTNAME: string;
  LASTNAME: string;
  CONFIRM_PASSWORD: string;
  MOBILENO: string;
  PASSWORD1: string;
  PASSWORD2: string;
  ADDRESS: string;
  ZIPCODE: string;
  CITY: string;
  COUNTRY: string;
  ACCOUNTNO: string;
  BILLING_EMAIL: string;
  FAX: string;
  ROLE: string;
  SITE: string;
  USERNAME_REQ: string;
  PASSWORD_REQ: string;
  ROLE_REQ: string;
  SITE_REQ: string;
  ENTITY_REQ: string;
  FIRSTNAME_REQ: string;
  LASTNAME_REQ: string;
  EMAIL_VALID: string;
  EMAIL_REQ: string;
  EMAIL_EXISTS: string;
  CONFIRMPASSWORD_REQ: string;
  MOBILE_REQ: string;
  ADDRESS_REQ: string;
  CITY_REQ: string;
  COUNTRY_REQ: string;
  ZIPCODE_REQ: string;
  BILLINGEMAIL_REQ: string;
  BILLING_EMAIL_VALID: string;
  BILLING_EMAIL_EXISTS: string;
  ACCOUNTNO_REQ: string;
  FAXNO_REQ: string;
  USER: string;
  TOKEN: string;
  FIRST_NAME: string;
  LAST_NAME: string;
  MOBILE_NO: string;
  ORGANIZATION_NAME: string;
  ORGANIZATIONNAME_REQ: string;
  ORG_NAME_EXISTS: string;
  CURRENT_PASSWORD: string;
  TRIAL: string;
  OWNER: string;
  LOGOUTSUCCESSION: string;
  REQUIRED_FIELD: string;
  NO_FILE: string;
}
// ...
export interface Buttons {
  DOWNLOAD: string;
  LOGIN: string;
  EDIT: string;
  REGISTER: string;
  SIGNUP: string;
  FORGOTPASSWORD: string;
  CANCEL: string;
  RESET: string;
  NEXT: string;
  BACK: string;
  SUBMIT: string;
  LOGINTEXT: string;
  RESEND_EMAIL: string;
  CHANGE_EMAIL: string;
  EDIT_ACCOUNT: string;
  CHANGE_PASSWORD: string;
  NO_THANKS: string;
  FIELDCOMMUNICATION: string;
  CHANGEPASSWORD: string;
  LEAVE: string;
  SAFETYBEAT: string;
  ENTRIES: string;
  ACTIVITIES: string;
  PROFIILE_INFO: string;
  JOIN_ENTITY: string;
  ADD_QUESTION: string;
  CREATE_QUESTION: string;
  CREATE_ENTITY: string;
  SKIP: string;
  MAP: string;
  HOURS: string;
  INVITE: string;
  INVITETEAM: string;
  ADD_SITE: string;
  IMPORT_SITE: string;
  SAVE: string;
  UPGRADE: string;
  GETSTARTED: string;
  LEARNMORE: string;
  DELETE: string;
  RENAME: string;
  INVITE_USER: string;
  UPLOAD_DOCUMENT: string;
  UPLOAD_FILE: string;
  VIEWPROFILE: string;
  CREATE_FOLDER: string;
  GENERATE_REPORT: string;
  VIEW_MAP: string;
  CLOSE: string;
  SHOW_RESULTS: string;
  PERMISSIONS: string;
  REPORTS: string;
  CHANGE_ACCESS_LEVEL: string;
  LOGOUT: string;
  HOW_IT_WORKS: string;
  ARCHIVED_HAZARDS: string;
  ARCHIVED_SITES: string;
  BUYPACKAGE: string;
  ARCHIVED_TEAMS: string;
  ARCHIVED_ENTITIES: string;
  CREATENEWQUESTION: string;
  ADDQUESTION: string;
  ADDUSER: string;
}

export interface Messages {
// ...
  REMOVE_CONNECTION_SUCCESS: string;
  REMOVE_CONNECTION_FAILURE: string;
  CONFIRM_CONNECTION_SUCCESS: string;
  CONFIRM_CONNECTION_FAILURE: string;
// ...
  ERROR_MSG: string;
  REMOVE_LEAVE_SUCCESS: string;
  REMOVE_LEAVE_FAILURE: string;
  CHECKIN_CATEGORIES: string;
  PULSE_CATEGORIES: string;
  CHECKIN_CATEGORY_TITLE: string;
  PULSE_CATEGORY_TITLE: string;
  BUYPACKAGE: string;
  PACKAGEUPDATE: string;
}

export interface Logger {
  STATUS: LoggerStatus;
  MESSAGES: LoggerMessage;
}

export interface LoggerStatus {
  DEFAULT: string;
  INFO: string;
  SUCCESS: string;
  WARNING: string;
  ERROR: string;
}

export interface LoggerMessage {
  LOGGEDIN: string;
  CREDENTIAL_REQ: string;
  TRUE: string;
  FALSE: string;
  STATUS: string;
  FORGOT_REQ: string;
  FORGOTSUCCESS: string;
  REGISTRATIONDATA_SUCCESS: string;
  REGISTRATION_REQ: string;
  REGISTRATION_SUCCESS: string;
  PROFILE_SUCCESS: string;
  PROFILE_ERROR: string;
  LOGIN_COMPONENT: string;
  REGISTRATION_COMPONENT: string;
  FORGOT_COMPONENT: string;
  PARTICLE_COMPONENT: string;
  DASHBOARD_COMPONENT: string;
  NAVIGATION_COMPONENT: string;
  SIDE_NAV: string;
  PROFILE_COMPONENT: string;
  VERIFICATION_COMPONENT: string;
  PROFILE_CREDENTIAL_REQ: string;
  PROFILE_UPDATED: string;
  PROFILE_NOTUPDATED: string;
  PASSWORD_CHANGE: string;
  PASSWORDCHANGE_UNSUCCESS: string;
  CHANGEPASSWORDFOR_DEV: string;
  ENTITYCONTROL: string;
  ORGANIZATIONDETAILS: string;
  CREATEENTITY: string;
  JOINENTITY: string;
  PASSWORDCHANGE: string;
  SETTING_COMPONENT: string;
  CREATEENTITY_ERROR: string;
  INVITEUSER_ERROR: string;
  ENTITYNOTCREATED: string;
  ROLES_RECIEVED: string;
  QUESTION_TYPES_RECEIVED: string;
  QUESTION_TYPES_RECEIVED_ERROR: string;
  ROLES_RECIEVED_ERROR: string;
  ORG_SUCCESS: string;
  QUESTION_DATA_REQ: string;
  ALL_QUESTION_RECEIVED: string;
  ALL_QUESTION_RECEIVED_ERROR: string;
}

export interface SiteTitle {
  MEMBERCENTER_TITLE: string;
  ENTITYCONTROL_TITLE: string;
  USERS_TITLE: string;
  PACKAGE_TITLE: string;
  SETTINGS_TITLE: string;
  SETTINGS_DESCRIPTION: string;
  SITE_CENTRE: string;
  QUESTION_CENTER: string;
  HAZARD_CENTER: string;
  DOCUMENTS: string;
  MY_TEAM: string;
  ACCOUNTINFO: string;
  ACTION_REPORT: string;
  AVERAGE_DAILY_ACTION_REPORT: string;
  NUMBER_CHECKINS_BY_ACTIVITY: string;
  HOURS_CHECKINS_BY_ACTIVITY: string;
  CHECKIN_BY_PERSON: string;
  ACTIONS_VS_ALERTS: string;
  PULSE_REPORT: string;
  COMPLIANT_CHECKOUT: string;
  SITE_ACTIVITY_REPORT: string;
  HAZARD_REPORTS: string;
  LEAVES: string;
  ADDLEAVES: string;
  MANAGELEAVES_TITLE: string;
}

export interface TableHeadings {
  ENTITY_ID: string;
  ENTITY_NAME: string;
  ENTITY_HEADOFFICE: string;
  ENTITY_STATUS: string;
  ACTIONS: string;
  NAME: string;
  EMAIL: string;
  CONTACT: string;
  ACCESS_LEVEL: string;
  SITE: string;
  LAST_ACTION: string;
  DATE_TIME: string;
  TIMEZONE: string;
  ADMINISTRATED_BY: string;
  SITE_NAME: string;
  RADIUS: string;
  RADIUSUNIT: string;
  SITE_ADDRESS: string;
  SITE_SAFE_ZONE: string;
  SITE_CREATED_BY: string;
  MANAGED_BY: string;
  SITE_SAFETY_MANAGER: string;
  NOTIFICATIONS: string;
  ENTITYCODE: string;
  REGISTER_TEAM: string;
  PAYMENT_HISTORY: string;
  DELETE_ENTITY: string;
  TITLE: string;
  HAZARD_DETAILS: string;
  TEAM_LEAD: string;
  ARCHIVE_ENTITY: string;
  AVERAGE_CHECK_IN: string;
  AVERAGE_CHECK_OUT: string;
  AVERAGE_DURATION: string;
}

export interface Status {
  DEFAULT: string;
  INFO: string;
  SUCCESS: string;
  WARNING: string;
  ERROR: string;
  CUSTOM: string;
}

export interface Confirmation {
  ACTIVATE_USER: string;
  DEACTIVATE_USER: string;
  UNSUSPEND_USER: string;
  SUSPEND_USER: string;
  DELETE_SITE: string;
  DELETE_HAZARD: string;
  DELETE_ENTITY: string;
  ADD_CONNECTION: string;
  CONFIRM_CONNECTION: string;
  DELETE_DOCUMENT: string;
  DELETE_FOLDER: string;
  REMOVE_CONNECTION: string
  DELETE_TEAM: string;
  DELETE_QUESTION: string;
  DELETE_LEAVE: string;
  ARCHIVE_SITE: string;
  ARCHIVE_HAZARD: string;
  ARCHIVE_ENTITY: string;
  ARCHIVE_TEAM: string;
  UNARCHIVE_TEAM: string;
}